refactor: refactor helpers tests and change file structure
