+++ src/view/components/Header/styles.module.scss
@@ -1,5 +1,5 @@
-@import '../../App/colors';
-@import '../../App/breakpoints';
+@import '../App/colors';
+@import '../App/breakpoints';
 
 .header {
   margin: 1.25em 0 1em;
